Move Walk/Run quick log to Workouts + menu
Relocate optional Walking/Running quick-log entry points from the Home
food + menu to the Workouts diary + menu on iOS and Android.

- Remove Home menu items, sheets, and ViewModel wiring
- Add gated menu items to WorkoutLogView / WorkoutDiaryScreen
- Move Settings toggle to Workout section with updated footer copy
- Reuse existing OutdoorActivityDurationSheet and logQuickCardio paths

Fixes #187
