+++ android/app/src/main/java/com/apoorvdarshan/calorietracker/ui/home/HomeScreen.kt
@@ -248,7 +248,6 @@ DisposableEffect(lifecycleOwner, vm) {
     val selectionMode = selectedFoodIds.isNotEmpty()
     var showNutritionDetail by remember { mutableStateOf(false) }
     var showCustomWaterLog by remember { mutableStateOf(false) }
-    var outdoorActivitySheet by remember { mutableStateOf<OutdoorActivityKind?>(null) }
     var showFastingStart by remember { mutableStateOf(false) }
     var editingFast by remember { mutableStateOf<FastingSession?>(null) }
     var pendingDiaryDeletion by remember { mutableStateOf<HomeDiaryItem?>(null) }
@@ -723,18 +722,6 @@ CalorieHero(
                                 SheetGlassDropdownMenuItem(label = stringResource(R.string.fasting), leadingIcon = Icons.Filled.Timer, trailingIcon = Icons.Filled.ChevronRight) { addMenuGroup = AddMenuGroup.Fasting }
                             }
                         }
-                        if (ui.walkRunQuickLogEnabled) {
-                            SheetGlassDropdownMenuItem(label = stringResource(R.string.outdoor_activity_walking), leadingIcon = Icons.AutoMirrored.Outlined.DirectionsWalk) {
-                                showAddMenu = false
-                                addMenuGroup = null
-                                outdoorActivitySheet = OutdoorActivityKind.WALKING
-                            }
-                            SheetGlassDropdownMenuItem(label = stringResource(R.string.outdoor_activity_running), leadingIcon = Icons.AutoMirrored.Filled.DirectionsRun) {
-                                showAddMenu = false
-                                addMenuGroup = null
-                                outdoorActivitySheet = OutdoorActivityKind.RUNNING
-                            }
-                        }
                     }
 
                     AddMenuGroup.PhotoAndScan -> {
@@ -805,14 +792,6 @@ CalorieHero(
             unit = ui.waterUnit,
             onDismiss = { showCustomWaterLog = false },
             onAdd = vm::addWater
-        )
-    }
-
-    outdoorActivitySheet?.let { activity ->
-        OutdoorActivityDurationSheet(
-            activity = activity,
-            onDismiss = { outdoorActivitySheet = null },
-            onLog = { minutes -> vm.logQuickOutdoorActivity(activity, minutes) }
         )
     }
