+++ android/app/src/main/java/com/apoorvdarshan/calorietracker/ui/settings/SettingsScreen.kt
@@ -888,13 +888,6 @@ fun SettingsScreen(container: AppContainer, nav: NavHostController, vm: Settings
                         icon = Icons.Outlined.TrackChanges
                     ) { sheet = SettingsSheet.FASTING_GOAL }
                 }
-                HorizontalDivider()
-                ToggleRow(
-                    stringResource(R.string.settings_walk_run_quick_log),
-                    ui.walkRunQuickLogEnabled,
-                    icon = Icons.AutoMirrored.Outlined.DirectionsWalk,
-                    onChange = vm::setWalkRunQuickLogEnabled
-                )
             }
             }
 
@@ -1247,6 +1240,20 @@ fun SettingsScreen(container: AppContainer, nav: NavHostController, vm: Settings
                 HorizontalDivider()
                 Text(
                     stringResource(R.string.settings_rpe_guide),
+                    fontSize = 12.sp,
+                    lineHeight = 18.sp,
+                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.55f),
+                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 10.dp)
+                )
+                HorizontalDivider()
+                ToggleRow(
+                    stringResource(R.string.settings_walk_run_quick_log),
+                    ui.walkRunQuickLogEnabled,
+                    icon = Icons.AutoMirrored.Outlined.DirectionsWalk,
+                    onChange = vm::setWalkRunQuickLogEnabled
+                )
+                Text(
+                    stringResource(R.string.settings_walk_run_quick_log_footer),
                     fontSize = 12.sp,
                     lineHeight = 18.sp,
                     color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.55f),
