+++ android/app/src/main/java/com/apoorvdarshan/calorietracker/ui/workouts/WorkoutDiaryScreen.kt
@@ -42,6 +42,8 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.text.TextAutoSize
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
+import androidx.compose.material.icons.automirrored.filled.DirectionsRun
+import androidx.compose.material.icons.automirrored.outlined.DirectionsWalk
 import androidx.compose.material.icons.filled.Mic
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Bookmark
@@ -66,6 +68,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.key
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateMapOf
@@ -118,6 +121,8 @@ import com.apoorvdarshan.calorietracker.models.WorkoutWeightUnit
 import com.apoorvdarshan.calorietracker.ui.components.FudGlassDialog
 import com.apoorvdarshan.calorietracker.ui.components.FudGlassSurface
 import com.apoorvdarshan.calorietracker.ui.components.FudGlassTextButton
+import com.apoorvdarshan.calorietracker.ui.home.OutdoorActivityDurationSheet
+import com.apoorvdarshan.calorietracker.ui.home.OutdoorActivityKind
 import com.apoorvdarshan.calorietracker.ui.home.SheetGlassDropdownMenu
 import com.apoorvdarshan.calorietracker.ui.home.SheetGlassDropdownMenuItem
 import com.apoorvdarshan.calorietracker.ui.navigation.BottomNavScrollPadding
@@ -157,6 +162,8 @@ internal fun WorkoutDiaryScreen(
     var workoutTranscript by androidx.compose.runtime.saveable.rememberSaveable { mutableStateOf("") }
     var copySheetVisible by remember { mutableStateOf(false) }
     var addMenuExpanded by remember { mutableStateOf(false) }
+    var outdoorActivitySheet by remember { mutableStateOf<OutdoorActivityKind?>(null) }
+    val walkRunQuickLogEnabled by container.prefs.walkRunQuickLogEnabled.collectAsState(initial = false)
     val listState = rememberLazyListState()
     val focusManager = LocalFocusManager.current
     val keyboard = LocalSoftwareKeyboardController.current
@@ -393,6 +400,24 @@ internal fun WorkoutDiaryScreen(
                         pickerRequest = WorkoutPickerRequest.saved()
                     }
                 )
+                if (walkRunQuickLogEnabled) {
+                    SheetGlassDropdownMenuItem(
+                        label = stringResource(R.string.outdoor_activity_walking),
+                        leadingIcon = Icons.AutoMirrored.Outlined.DirectionsWalk,
+                        onClick = {
+                            addMenuExpanded = false
+                            outdoorActivitySheet = OutdoorActivityKind.WALKING
+                        }
+                    )
+                    SheetGlassDropdownMenuItem(
+                        label = stringResource(R.string.outdoor_activity_running),
+                        leadingIcon = Icons.AutoMirrored.Filled.DirectionsRun,
+                        onClick = {
+                            addMenuExpanded = false
+                            outdoorActivitySheet = OutdoorActivityKind.RUNNING
+                        }
+                    )
+                }
                 HorizontalDivider(color = workoutsColors().hairline.copy(alpha = 0.45f))
                 SheetGlassDropdownMenuItem(label = stringResource(R.string.workout_text_voice), leadingIcon = Icons.Filled.Mic, onClick = {
                     addMenuExpanded = false
@@ -451,6 +476,16 @@ internal fun WorkoutDiaryScreen(
             onToggleExercise = viewModel::toggleExercise,
             onToggleSaved = viewModel::toggleSaved,
             onDismiss = { pickerRequest = null }
+        )
+    }
+
+    outdoorActivitySheet?.let { activity ->
+        OutdoorActivityDurationSheet(
+            activity = activity,
+            onDismiss = { outdoorActivitySheet = null },
+            onLog = { minutes ->
+                viewModel.logQuickOutdoorActivity(activity, minutes)
+            }
         )
     }
